+++ plotnine/coords/coord.py
@@ -112,8 +112,8 @@ class coord(object):
         """
         return data
 
-    def expand_default(self, scale, discrete=(0, 0.6),
-                       continuous=(0.05, 0)):
+    def expand_default(self, scale, discrete=(0, 0.6, 0, 0.6),
+                       continuous=(0.05, 0, 0.05, 0)):
         """
         Expand a single scale
         """